fix(runtime): tolerate codex exec progress items
